Bug 1314120 - Socorro Import - Add schema generation and versioning in s3 path (#229)
* Bug 1314120 - Generate Spark schema from JSON schema

This replaces the hardcoded schema with one that is generated from the
JSON schema that is published alongside the crash data. This also
replaces references to print to logging statements.

* Bug 1314120 - Add version to output path

This uses the schema `$target_version` as part of the path for the
output bucket. This defaults to v0 for schemas that do not include a
versioning system.

* Bug 1314120 - Add boto and better exceptions

This modifies downloading the JSON schema from an s3 bucket using boto.
This adds more descriptive exceptions among other small fixes.

* Bug 1314120 - Add region to s3 client

* Bug 1314120 - Update source bucket to production

We also update the destination bucket to a new testing location.

* Bug 1314120 - Update number of parquet partitions

Increase the number of partitions from 5 to 10 to reduce the size of the
partitions in a range roughly between 300 MB and 1 GB.

* Bug 1314120 - Isolate operational code to one cell

ATMO currently continues the execution of a notebook when there are
errors. This commit isolates the main portion of the code into a single
cell.

* Bug 1314120 - Change destination to production bucket

This finally changes the destination bucket to the final
telemetry-parquet bucket.