Bug 1319592 Use correct latest version
Look up the major release version history on
product-details.mozilla.org to find what version was current as of
the beginning of the week under consideration, instead of pulling
version.txt from hg which can be updated ahead of time.

This also allows the job to be safely re-run for old dates.